# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CMMS API — Beit Jala Pharmaceutical Co.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signatures:
    get:
      operationId: listSignatures
      tags: [signatures]
      summary: List immutable signatures for a document
      parameters:
        - name: documentType
          in: query
          required: true
          schema:
            type: string
        - name: documentId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signatures for the document
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signature"
...
  /maintenance-requests/closed-log/manual:
    post:
      operationId: createManualClosedCorrectiveMaintenanceLogEntry
      tags: [maintenance-requests]
      summary: Add a supplemental manual row to LOG-10-0659-0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualClosedCorrectiveMaintenanceLogEntryInput"
      responses:
        "201":
          description: Manual log entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClosedCorrectiveMaintenanceLogRow"
...
components:
...
  schemas:
...
    Signature:
      type: object
      required:
        [
          id,
          documentType,
          documentId,
          fieldName,
          signatureType,
          userId,
          userName,
          signedAt,
        ]
      properties:
        id:
          type: integer
        documentType:
          type: string
        documentId:
          type: integer
        fieldName:
          type: string
        signatureType:
          type: string
        userId:
          type: integer
        userName:
          type: string
        signatureData:
          type: ["string", "null"]
        eligibleSignerAssignmentId:
          type: ["integer", "null"]
        signedAt:
          type: string
...
    ClosedCorrectiveMaintenanceLogRow:
      type: object
      required:
        [
          id,
          source,
          machineName,
          machineNumber,
          requestDate,
          requestReportNumber,
          priority,
          closedDate,
          remarks,
        ]
      properties:
        id:
          type: string
        source:
          type: string
          enum: [automatic, manual, corrective-record]
        machineName:
          type: string
        machineNumber:
          type: string
        requestDate:
          type: string
        requestReportNumber:
          type: string
        priority:
          type: string
        closedDate:
          type: string
        remarks:
          type: string
    ManualClosedCorrectiveMaintenanceLogEntryInput:
      type: object
      required:
        [
          machineName,
          machineNumber,
          requestDate,
          requestReportNumber,
          closedDate,
        ]
      properties:
        machineName:
          type: string
        machineNumber:
          type: string
        requestDate:
          type: string
        requestReportNumber:
          type: string
        priority:
          type: string
          enum: [normal, urgent]
        closedDate:
          type: string
        remarks:
          type: string
...